#1539 Create common parent for SQLite and MySQL
