+++ src/main/java/fr/xephi/authme/datasource/MySQL.java
@@ -1,19 +1,13 @@
 package fr.xephi.authme.datasource;
 
-import ch.jalu.datasourcecolumns.data.DataSourceValue;
-import ch.jalu.datasourcecolumns.data.DataSourceValueImpl;
-import ch.jalu.datasourcecolumns.data.DataSourceValues;
-import ch.jalu.datasourcecolumns.predicate.AlwaysTruePredicate;
 import com.google.common.annotations.VisibleForTesting;
 import com.zaxxer.hikari.HikariDataSource;
 import com.zaxxer.hikari.pool.HikariPool.PoolInitializationException;
 import fr.xephi.authme.ConsoleLogger;
 import fr.xephi.authme.data.auth.PlayerAuth;
-import fr.xephi.authme.datasource.columnshandler.AuthMeColumns;
 import fr.xephi.authme.datasource.columnshandler.AuthMeColumnsHandler;
 import fr.xephi.authme.datasource.mysqlextensions.MySqlExtension;
 import fr.xephi.authme.datasource.mysqlextensions.MySqlExtensionsFactory;
-import fr.xephi.authme.security.crypts.HashedPassword;
 import fr.xephi.authme.settings.Settings;
 import fr.xephi.authme.settings.properties.DatabaseSettings;
 import fr.xephi.authme.settings.properties.HooksSettings;
@@ -26,14 +20,10 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import static ch.jalu.datasourcecolumns.data.UpdateValues.with;
-import static ch.jalu.datasourcecolumns.predicate.StandardPredicates.eq;
-import static ch.jalu.datasourcecolumns.predicate.StandardPredicates.eqIgnoreCase;
 import static fr.xephi.authme.datasource.SqlDataSourceUtils.getNullableLong;
 import static fr.xephi.authme.datasource.SqlDataSourceUtils.logSqlException;
 
@@ -41,7 +31,7 @@ import static fr.xephi.authme.datasource.SqlDataSourceUtils.logSqlException;
  * MySQL data source.
  */
 @SuppressWarnings({"checkstyle:AbbreviationAsWordInName"}) // Justification: Class name cannot be changed anymore
-public class MySQL implements DataSource {
+public class MySQL extends AbstractSqlDataSource {
 
     private boolean useSsl;
     private String host;
@@ -54,7 +44,6 @@ public class MySQL implements DataSource {
     private int maxLifetime;
     private List<String> columnOthers;
     private Columns col;
-    private AuthMeColumnsHandler columnsHandler;
     private MySqlExtension sqlExtension;
     private HikariDataSource ds;
 
@@ -274,29 +263,6 @@ public class MySQL implements DataSource {
     }
 
     @Override
-    public boolean isAuthAvailable(String user) {
-        try {
-            return columnsHandler.retrieve(user, AuthMeColumns.NAME).rowExists();
-        } catch (SQLException e) {
-            logSqlException(e);
-            return false;
-        }
-    }
-
-    @Override
-    public HashedPassword getPassword(String user) {
-        try {
-            DataSourceValues values = columnsHandler.retrieve(user, AuthMeColumns.PASSWORD, AuthMeColumns.SALT);
-            if (values.rowExists()) {
-                return new HashedPassword(values.get(AuthMeColumns.PASSWORD), values.get(AuthMeColumns.SALT));
-            }
-        } catch (SQLException e) {
-            logSqlException(e);
-        }
-        return null;
-    }
-
-    @Override
     public PlayerAuth getAuth(String user) {
         String sql = "SELECT * FROM " + tableName + " WHERE " + col.NAME + "=?;";
         PlayerAuth auth;
@@ -318,9 +284,7 @@ public class MySQL implements DataSource {
 
     @Override
     public boolean saveAuth(PlayerAuth auth) {
-        columnsHandler.insert(auth,
-            AuthMeColumns.NAME, AuthMeColumns.NICK_NAME, AuthMeColumns.PASSWORD, AuthMeColumns.SALT,
-            AuthMeColumns.EMAIL, AuthMeColumns.REGISTRATION_DATE, AuthMeColumns.REGISTRATION_IP);
+        super.saveAuth(auth);
 
         try (Connection con = getConnection()) {
             if (!columnOthers.isEmpty()) {
@@ -340,23 +304,6 @@ public class MySQL implements DataSource {
             logSqlException(ex);
         }
         return false;
-    }
-
-    @Override
-    public boolean updatePassword(PlayerAuth auth) {
-        return updatePassword(auth.getNickname(), auth.getPassword());
-    }
-
-    @Override
-    public boolean updatePassword(String user, HashedPassword password) {
-        return columnsHandler.update(user,
-            with(AuthMeColumns.PASSWORD, password.getHash())
-            .and(AuthMeColumns.SALT, password.getSalt()).build());
-    }
-
-    @Override
-    public boolean updateSession(PlayerAuth auth) {
-        return columnsHandler.update(auth, AuthMeColumns.LAST_IP, AuthMeColumns.LAST_LOGIN, AuthMeColumns.NICK_NAME);
     }
 
     @Override
@@ -397,37 +344,10 @@ public class MySQL implements DataSource {
     }
 
     @Override
-    public boolean updateQuitLoc(PlayerAuth auth) {
-        return columnsHandler.update(auth,
-            AuthMeColumns.LOCATION_X, AuthMeColumns.LOCATION_Y, AuthMeColumns.LOCATION_Z,
-            AuthMeColumns.LOCATION_WORLD, AuthMeColumns.LOCATION_YAW, AuthMeColumns.LOCATION_PITCH);
-    }
-
-    @Override
-    public boolean updateEmail(PlayerAuth auth) {
-        return columnsHandler.update(auth, AuthMeColumns.EMAIL);
-    }
-
-    @Override
     public void closeConnection() {
         if (ds != null && !ds.isClosed()) {
             ds.close();
         }
-    }
-
-    @Override
-    public List<String> getAllAuthsByIp(String ip) {
-        try {
-            return columnsHandler.retrieve(eq(AuthMeColumns.LAST_IP, ip), AuthMeColumns.NAME);
-        } catch (SQLException e) {
-            logSqlException(e);
-            return Collections.emptyList();
-        }
-    }
-
-    @Override
-    public int countAuthsByEmail(String email) {
-        return columnsHandler.count(eqIgnoreCase(AuthMeColumns.EMAIL, email));
     }
 
     @Override
@@ -446,73 +366,6 @@ public class MySQL implements DataSource {
     @Override
     public DataSourceType getType() {
         return DataSourceType.MYSQL;
-    }
-
-    @Override
-    public boolean isLogged(String user) {
-        try {
-            DataSourceValue<Integer> result = columnsHandler.retrieve(user, AuthMeColumns.IS_LOGGED);
-            return result.rowExists() && Integer.valueOf(1).equals(result.getValue());
-        } catch (SQLException e) {
-            logSqlException(e);
-            return false;
-        }
-    }
-
-    @Override
-    public void setLogged(String user) {
-        columnsHandler.update(user, AuthMeColumns.IS_LOGGED, 1);
-    }
-
-    @Override
-    public void setUnlogged(String user) {
-        columnsHandler.update(user, AuthMeColumns.IS_LOGGED, 0);
-    }
-
-    @Override
-    public boolean hasSession(String user) {
-        try {
-            DataSourceValue<Integer> result = columnsHandler.retrieve(user, AuthMeColumns.HAS_SESSION);
-            return result.rowExists() && Integer.valueOf(1).equals(result.getValue());
-        } catch (SQLException e) {
-            logSqlException(e);
-            return false;
-        }
-    }
-
-    @Override
-    public void grantSession(String user) {
-        columnsHandler.update(user, AuthMeColumns.HAS_SESSION, 1);
-    }
-
-    @Override
-    public void revokeSession(String user) {
-        columnsHandler.update(user, AuthMeColumns.HAS_SESSION, 0);
-    }
-
-    @Override
-    public void purgeLogged() {
-        columnsHandler.update(eq(AuthMeColumns.IS_LOGGED, 1), AuthMeColumns.IS_LOGGED, 0);
-    }
-
-    @Override
-    public int getAccountsRegistered() {
-        return columnsHandler.count(new AlwaysTruePredicate<>());
-    }
-
-    @Override
-    public boolean updateRealName(String user, String realName) {
-        return columnsHandler.update(user, AuthMeColumns.NICK_NAME, realName);
-    }
-
-    @Override
-    public DataSourceValue<String> getEmail(String user) {
-        try {
-            return columnsHandler.retrieve(user, AuthMeColumns.EMAIL);
-        } catch (SQLException e) {
-            logSqlException(e);
-            return DataSourceValueImpl.unknownRow();
-        }
     }
 
     @Override
