+++ src/test/java/fr/xephi/authme/ClassesConsistencyTest.java
@@ -5,6 +5,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import fr.xephi.authme.data.captcha.CaptchaCodeStorage;
+import fr.xephi.authme.datasource.AbstractSqlDataSource;
 import fr.xephi.authme.datasource.Columns;
 import fr.xephi.authme.datasource.columnshandler.DataSourceColumn;
 import fr.xephi.authme.datasource.columnshandler.PlayerAuthColumn;
@@ -62,6 +63,7 @@ public class ClassesConsistencyTest {
     private static final Set<Class<?>> CLASSES_EXCLUDED_FROM_VISIBILITY_TEST = ImmutableSet.of(
         Whirlpool.class, // not our implementation, so we don't touch it
         MySqlExtension.class, // has immutable protected fields used by all children
+        AbstractSqlDataSource.class, // protected members for inheritance
         Columns.class // uses non-static String constants, which is safe
     );
